+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/riva/repos/todo/src/reportWebVitals.js":"1","/Users/riva/repos/todo/src/App.js":"2","/Users/riva/repos/todo/src/index.js":"3","/Users/riva/repos/todo/src/Components/TodoList.js":"4","/Users/riva/repos/todo/src/Components/TodoForm.js":"5","/Users/riva/repos/todo/src/Components/ToDoConnected.js":"6","/Users/riva/repos/todo/src/Components/Todo.js":"7"},{"size":362,"mtime":1606171492211,"results":"8","hashOfConfig":"9"},{"size":379,"mtime":1606186167534,"results":"10","hashOfConfig":"9"},{"size":336,"mtime":1606186167795,"results":"11","hashOfConfig":"9"},{"size":583,"mtime":1606186167756,"results":"12","hashOfConfig":"9"},{"size":1284,"mtime":1606185687635,"results":"13","hashOfConfig":"9"},{"size":1958,"mtime":1606182754195,"results":"14","hashOfConfig":"9"},{"size":1914,"mtime":1606186167590,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"czx635",{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":"20","usedDeprecatedRules":"21"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/riva/repos/todo/src/reportWebVitals.js",[],"/Users/riva/repos/todo/src/App.js",["32","33"],"import './App.css';\nimport React, { useState } from 'react';\nimport Todo from './Components/Todo';\n\n\nconst App = () => {\n  const [ input, countInput ] = useState(0);\n\n  return (\n    <div className=\"App\">\n      <div className=\"App-header\">\n        <div>Home </div>\n      </div>\n      <div className=\"formAndTodo\">\n      <Todo/>\n      </div>\n    </div>\n  );\n}\n\nexport default App;\n",["34","35"],"/Users/riva/repos/todo/src/index.js",[],"/Users/riva/repos/todo/src/Components/TodoList.js",[],"/Users/riva/repos/todo/src/Components/TodoForm.js",["36","37","38"],"/Users/riva/repos/todo/src/Components/ToDoConnected.js",[],"/Users/riva/repos/todo/src/Components/Todo.js",[],{"ruleId":"39","severity":1,"message":"40","line":7,"column":11,"nodeType":"41","messageId":"42","endLine":7,"endColumn":16},{"ruleId":"39","severity":1,"message":"43","line":7,"column":18,"nodeType":"41","messageId":"42","endLine":7,"endColumn":28},{"ruleId":"44","replacedBy":"45"},{"ruleId":"46","replacedBy":"47"},{"ruleId":"39","severity":1,"message":"48","line":1,"column":27,"nodeType":"41","messageId":"42","endLine":1,"endColumn":36},{"ruleId":"39","severity":1,"message":"49","line":9,"column":9,"nodeType":"41","messageId":"42","endLine":9,"endColumn":26},{"ruleId":"39","severity":1,"message":"50","line":13,"column":9,"nodeType":"41","messageId":"42","endLine":13,"endColumn":21},"no-unused-vars","'input' is assigned a value but never used.","Identifier","unusedVar","'countInput' is assigned a value but never used.","no-native-reassign",["51"],"no-negated-in-lhs",["52"],"'useEffect' is defined but never used.","'handleInputChange' is assigned a value but never used.","'handleSubmit' is assigned a value but never used.","no-global-assign","no-unsafe-negation"]
+[{"/Users/riva/repos/todo/src/reportWebVitals.js":"1","/Users/riva/repos/todo/src/App.js":"2","/Users/riva/repos/todo/src/index.js":"3","/Users/riva/repos/todo/src/Components/TodoList.js":"4","/Users/riva/repos/todo/src/Components/TodoForm.js":"5","/Users/riva/repos/todo/src/Components/ToDoConnected.js":"6","/Users/riva/repos/todo/src/Components/Todo.js":"7","/Users/riva/repos/todo/src/useFormHook.js":"8"},{"size":362,"mtime":1606171492211,"results":"9","hashOfConfig":"10"},{"size":379,"mtime":1606186167534,"results":"11","hashOfConfig":"10"},{"size":335,"mtime":1606197327577,"results":"12","hashOfConfig":"10"},{"size":583,"mtime":1606186167756,"results":"13","hashOfConfig":"10"},{"size":1103,"mtime":1606260168371,"results":"14","hashOfConfig":"10"},{"size":1958,"mtime":1606182754195,"results":"15","hashOfConfig":"10"},{"size":1913,"mtime":1606197248986,"results":"16","hashOfConfig":"10"},{"size":544,"mtime":1606259927299,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"czx635",{"filePath":"20","messages":"21","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":"22","usedDeprecatedRules":"23"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":4,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/riva/repos/todo/src/reportWebVitals.js",[],"/Users/riva/repos/todo/src/App.js",["36","37"],"import './App.css';\nimport React, { useState } from 'react';\nimport Todo from './Components/Todo';\n\n\nconst App = () => {\n  const [ input, countInput ] = useState(0);\n\n  return (\n    <div className=\"App\">\n      <div className=\"App-header\">\n        <div>Home </div>\n      </div>\n      <div className=\"formAndTodo\">\n      <Todo/>\n      </div>\n    </div>\n  );\n}\n\nexport default App;\n",["38","39"],"/Users/riva/repos/todo/src/index.js",[],"/Users/riva/repos/todo/src/Components/TodoList.js",[],"/Users/riva/repos/todo/src/Components/TodoForm.js",["40","41","42","43"],"/Users/riva/repos/todo/src/Components/ToDoConnected.js",[],"/Users/riva/repos/todo/src/Components/Todo.js",[],"/Users/riva/repos/todo/src/useFormHook.js",["44"],{"ruleId":"45","severity":1,"message":"46","line":7,"column":11,"nodeType":"47","messageId":"48","endLine":7,"endColumn":16},{"ruleId":"45","severity":1,"message":"49","line":7,"column":18,"nodeType":"47","messageId":"48","endLine":7,"endColumn":28},{"ruleId":"50","replacedBy":"51"},{"ruleId":"52","replacedBy":"53"},{"ruleId":"45","severity":1,"message":"54","line":1,"column":17,"nodeType":"47","messageId":"48","endLine":1,"endColumn":25},{"ruleId":"45","severity":1,"message":"55","line":1,"column":27,"nodeType":"47","messageId":"48","endLine":1,"endColumn":36},{"ruleId":"45","severity":1,"message":"56","line":8,"column":13,"nodeType":"47","messageId":"48","endLine":8,"endColumn":19},{"ruleId":"45","severity":1,"message":"57","line":8,"column":21,"nodeType":"47","messageId":"48","endLine":8,"endColumn":33},{"ruleId":"45","severity":1,"message":"58","line":1,"column":8,"nodeType":"47","messageId":"48","endLine":1,"endColumn":13},"no-unused-vars","'input' is assigned a value but never used.","Identifier","unusedVar","'countInput' is assigned a value but never used.","no-native-reassign",["59"],"no-negated-in-lhs",["60"],"'useState' is defined but never used.","'useEffect' is defined but never used.","'values' is assigned a value but never used.","'handleChange' is assigned a value but never used.","'React' is defined but never used.","no-global-assign","no-unsafe-negation"]
